pull workflow published buttons into separate component
