+++ client/src/components/Workflow/Published/WorkflowPublished.vue
@@ -4,7 +4,7 @@ import { faBuilding, faDownload, faEdit, faPlay, faSpinner, faUser } from "@fort
 import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
 import { until } from "@vueuse/core";
 import { type AxiosError } from "axios";
-import { BAlert, BButton, BCard } from "bootstrap-vue";
+import { BAlert, BCard } from "bootstrap-vue";
 import { computed, nextTick, onMounted, ref, watch } from "vue";
 
 import { getWorkflowInfo, type StoredWorkflowDetailed } from "@/api/workflows";
@@ -12,15 +12,14 @@ import { fromSimple } from "@/components/Workflow/Editor/modules/model";
 import { getWorkflowFull } from "@/components/Workflow/workflows.services";
 import { useDatatypesMapper } from "@/composables/datatypesMapper";
 import { provideScopedWorkflowStores } from "@/composables/workflowStores";
-import { useUserStore } from "@/stores/userStore";
 import type { Steps } from "@/stores/workflowStepStore";
 import { assertDefined } from "@/utils/assertions";
-import { withPrefix } from "@/utils/redirect";
 
 import ActivityBar from "@/components/ActivityBar/ActivityBar.vue";
 import Heading from "@/components/Common/Heading.vue";
 import WorkflowGraph from "@/components/Workflow/Editor/WorkflowGraph.vue";
 import WorkflowInformation from "@/components/Workflow/Published/WorkflowInformation.vue";
+import WorkflowPublishedButtons from "@/components/Workflow/Published/WorkflowPublishedButtons.vue";
 
 library.add(faBuilding, faDownload, faEdit, faPlay, faSpinner, faUser);
 
@@ -51,8 +50,6 @@ const props = withDefaults(defineProps<Props>(), {
     showZoomControls: true,
 });
 
-const userStore = useUserStore();
-
 const { datatypesMapper } = useDatatypesMapper();
 
 const { stateStore } = provideScopedWorkflowStores(props.id);
@@ -64,43 +61,13 @@ const workflow = ref<StoredWorkflowDetailed | null>(null);
 
 const hasError = computed(() => !!errorMessage.value);
 
-const downloadUrl = computed(() => withPrefix(`/api/workflows/${props.id}/download?format=json-download`));
-const importUrl = computed(() => withPrefix(`/workflow/imp?id=${props.id}`));
-const runUrl = computed(() => withPrefix(`/workflows/run?id=${props.id}`));
-
 const initialPosition = computed(() => ({
     x: -props.initialX * props.zoom,
     y: -props.initialY * props.zoom,
 }));
 
-const viewUrl = computed(() => withPrefix(`/published/workflow?id=${props.id}`));
-
-const sharedWorkflow = computed(() => {
-    return !userStore.matchesCurrentUsername(workflowInfo.value?.owner);
-});
-
-const editButtonTitle = computed(() => {
-    if (userStore.isAnonymous) {
-        return "Log in to edit Workflow";
-    } else {
-        if (workflowInfo.value?.deleted) {
-            return "You cannot edit a deleted workflow. Restore it first.";
-        } else {
-            return "Edit Workflow";
-        }
-    }
-});
-
 /** Workflow steps force typed as `Steps` from the `workflowStepStore` */
 const workflowSteps = computed(() => (workflow.value?.steps as unknown as Steps) ?? []);
-
-function logInTitle(title: string) {
-    if (userStore.isAnonymous) {
-        return `Log in to ${title}`;
-    } else {
-        return title;
-    }
-}
 
 async function load() {
     errorMessage.value = "";
@@ -144,6 +111,11 @@ onMounted(async () => {
     // @ts-ignore: TS2339 webpack dev issue. hopefully we can remove this with vite
     workflowGraph.value?.fitWorkflow(0.25, 1.5, 20.0);
 });
+
+defineExpose({
+    workflow,
+    workflowInfo,
+});
 </script>
 
 <template>
@@ -171,65 +143,11 @@ onMounted(async () => {
                         <span v-else> {{ workflowInfo.name }} </span>
                     </Heading>
 
-                    <span v-if="props.showButtons">
-                        <BButton
-                            v-b-tooltip.hover.noninteractive
-                            title="Download workflow in .ga format"
-                            variant="outline-primary"
-                            size="md"
-                            :href="downloadUrl">
-                            <FontAwesomeIcon :icon="faDownload" />
-                            Download
-                        </BButton>
-
-                        <BButton
-                            v-if="!props.embed && sharedWorkflow"
-                            :href="importUrl"
-                            :disabled="userStore.isAnonymous"
-                            :title="logInTitle('Import Workflow')"
-                            data-description="workflow import"
-                            target="_blank"
-                            variant="outline-primary"
-                            size="md">
-                            <FontAwesomeIcon :icon="faEdit" />
-                            Import
-                        </BButton>
-
-                        <BButton
-                            v-else-if="!props.embed && !sharedWorkflow"
-                            v-b-tooltip.hover.noninteractive
-                            :disabled="workflowInfo.deleted"
-                            class="workflow-edit-button"
-                            :title="editButtonTitle"
-                            variant="outline-primary"
-                            size="md"
-                            :to="`/workflows/edit?id=${workflowInfo.id}`">
-                            <FontAwesomeIcon :icon="faEdit" fixed-width />
-                            Edit
-                        </BButton>
-
-                        <BButton
-                            v-if="!props.embed"
-                            :to="runUrl"
-                            :disabled="userStore.isAnonymous"
-                            :title="logInTitle('Run Workflow')"
-                            variant="primary"
-                            size="md">
-                            <FontAwesomeIcon :icon="faPlay" />
-                            Run
-                        </BButton>
-
-                        <BButton
-                            v-if="props.embed"
-                            :href="viewUrl"
-                            target="blank"
-                            variant="primary"
-                            size="md"
-                            class="view-button font-weight-bold">
-                            <FontAwesomeIcon :icon="['gxd', 'galaxyLogo']" />
-                            View In Galaxy
-                        </BButton>
-                    </span>
+                    <WorkflowPublishedButtons
+                        v-if="props.showButtons"
+                        :id="props.id"
+                        :embed="props.embed"
+                        :workflow-info="workflowInfo" />
                 </div>
 
                 <BCard class="workflow-preview" :class="{ 'only-preview': !props.showAbout }">
